% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_models.R
\name{build_models}
\alias{build_models}
\title{Build models of storm characteristics from existing data.}
\usage{
build_models(storms = NULL, strip = FALSE)
}
\arguments{
\item{storms}{A list of storm tabular data as output by the \code{get_storms()} function. Ideally, this should include the largest possible pool of storms relevant for
subsequent analyses, as the models will reflect variability in these input storms only.}

\item{strip}{Whether to strip down the models and the nested structure into separate lists, rather than nested data sets. This helps when trying to view the
models or the underlying data or when trying to save to disk.}
}
\value{
A list of nested tibbles, one for each of the model groups:
\itemize{
\item distmods:  asymptotic and linear models of distance from the storm center as a function of wind speed. Models nested by basin, quadrant, and Saffir-Simpson category.
\item emod:      linear models of eye radius as a function of the radius of maximum winds. Models nested by basin and Saffir-Simpson category.
\item rocimod:   linear models of the radius of the last closed isobar (ROCI: the storm's maximum extent) as a function of the maximum radius of windspeeds.
Models nested by basin and Saffir-Simpson category.
\item pocimod:   linear models of the pressure differential (the minimum pressure at the center and the atmospheric pressure at the ROCI) as a function of the maximum wind speed.
Models nested by basin and Saffir-Simpson category.
}

For all models, variable units are native to those of USA IBTrACS variables: speed is windspeed in knots, dist is distance from center in nautical miles, and pressure is millibars.
}
\description{
\code{build_models()} will accept the output of \code{get_storms()} and create a series of statistical models relating storm characteristics.
This includes the relationship between a given wind speed and its distance from the center, the eyewall radius as a function of the radius of maximum wind,
the pressure differential in a storm as a function of the maximum wind speed, & the radius of the last closed isobar as a function of the maximum radius of
given wind speeds. Models are nested by basin, Saffir-Simpson intensity category, and by storm quadrant. Models will only reflect the variability of input
storms. If input storms were previously consolidated, those values will be modeled, otherwise all values are modeled.

These models are currently only used in the \code{make_extents()} function to fill in missing data when necessary, primarily to build wind extents for older storms.
}
\examples{

#  build models off all IBTrACS storms
allstorms <- get_storms(ib_filt="ALL")
mods <- build_models(allstorms)

# plot training and prediction data
distmoddat <- mods$distmods |> unnest(data)
samp <- sample(seq(1:nrow(distmoddat)),1000)
library(ggplot2)
ggplot(data=distmoddat[samp,],aes(x=windspeed,y=dist,col=factor(USA_SSHS),shape=BASIN))+geom_point()+geom_line(aes(y=pred_asymp.dist))+geom_line(aes(y=pred_lm.dist),lty=2)+facet_wrap(~BASIN)

}
