\name{etopo60}
\alias{etopo60}
\alias{data(etopo60)}
\title{Topography with 60-minute resolution. This is a low-resolution
  version of the more recent 2-minute topography: ETOPO5}
\description{.}
\usage{data(etopo60) }
\format{list: ETOPO60X, ETOPO60Y, ROSE}
\source{\url{http://www.ngdc.noaa.gov/mgg/global/etopo5.HTML} and \url{}}
\references{\url{http://www.ngdc.noaa.gov/mgg/bathymetry/predicted/explore.HTML}}
\examples{
data(etopo60)
}
\keyword{datasets}


