% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surge.R
\name{get_surge}
\alias{get_surge}
\title{Generate 2D storm surge rasters for a given storm}
\usage{
get_surge(
  storm,
  dpath = NULL,
  todir = NULL,
  cpus = NULL,
  overwrite = FALSE,
  returnpoints = TRUE,
  usgskey = NULL
)
}
\arguments{
\item{storm}{A set of storm tabular data generated by 'get_storms()'. If not used with an apply function, only the first storm will be processed.}

\item{dpath}{The download path of the sourced elevation and water level data. Raw files will be stored here and then searched for here before
downloading from the web. An "./Elevation/" and "./Water Levels/" directory will be created for each group of data.}

\item{cpus}{The number of cpus to use in parallel processing.}
}
\description{
\code{get_surge()} will accept the output of 'get_storms()' and return a single layer raster representing the maximum storm surge during the life of the storm.
Storm surge is a product of surface elevation and water levels. It is calculated as the maximum water depth, in meters, over land only where the elevation is
greater than the mean high tide from the previous year. Surface elevation comes from NASA's Shuttle Radar Topography Mission (SRTM), which
has global coverage at 30 m resolution. These will be applied to all storms and only need to be downloaded once per location. They may also be used
for wind topography interactions in future package versions. Water levels come from three sources and these sources mostly determine the quality of
the final surge product. In coastal areas of the United States, data from the National Oceanic and Atmospheric Administration (NOAA) and the United States
Geological Survey (USGS) provide very accurate observations with good geographic coverage for relatively recent storms.
The European Union's Coupled Model Intercomparison Project (CMIP6) provides global coverage back to 1850 but as an ensemble model product, the values
are less accurate that those from NOAA and USGS. All three sources are used to interpolate water levels between observation locations, but preference is
given to usgs and noaa values when available.
The spatial resolution of the surge raster, for now, is limited to that of the elevation data.
}
\examples{
Sandy <- get_storms(name='Sandy',season=2012)
Sandy_surge <- get_surge(Sandy,dpath="I:/")
###  focus on New York
Sandy_surge_NYC <- crop(Sandy_surge,ext(c(-75,-73.5,40,41)))
plot(Sandy_surge_NYC)

##  A non-USA/historical storm
##  will rely entirely on CMIP6 model output
Nargis <- get_storms(name='Nargis',season=2008)
Nargis_surge <- get_surge(Nargis,dpath="I:/")
Nargis_surge_Myanmar <- crop(Nargis_surge,ext(c(90,98,13.5,19)))
plot(Nargis_surge_Myanmar)
}
