% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wind.R
\name{get_wind}
\alias{get_wind}
\title{Generate 2D wind fields for a storm}
\usage{
get_wind(
  stormextent,
  s_res = 20000,
  methods = NULL,
  cpus = NULL,
  todir = NULL,
  overwrite = FALSE,
  smooth = FALSE,
  eye_option = "maxwind",
  loadrasts = FALSE,
  agg = FALSE
)
}
\arguments{
\item{stormextent}{A set of storm wind extents generated by the 'make_extents()' function. If not used with an apply function, only the first storm will be processed.}

\item{s_res}{The spatial resolution, in meters, of the desired output. Higher resolutions (lower numbers) will take significantly longer and parallel processing may be warranted below
10000 meters.}

\item{methods}{One or all of 'tps', 'boose', 'willoughby', or 'holland'. See details below. 'all' will perform all of the methods.}

\item{cpus}{How many cpu cores to assign for parallel processing. Recommended to verify available cores before using.}

\item{todir}{The target directory for final rasters. If not provided, rasters will be loaded in memory. Thus, highly recommended for high resolution products or for batch processing.}

\item{overwrite}{Should existing files in 'todir' be overwritten. If not, those will be skipped.}

\item{smooth}{Should a smoothing function be applied, analagous to that applied to StormR products.}

\item{eye_option}{How the eye wall is processed when creating rasters. See details.}
}
\value{
Raster stacks, one for each time step in the output of 'make_extents()', with the spatial resolution of 's_res' and the spatial extent of the cumulative radius of the last
closed isobar (ROCI), representing the maximum sustained wind in meters per second, the wind direction in degrees, and the total power in kilowatts exerted by the maximum sustained
wind on 1 square meter of a surface for the duration of the time step.
}
\description{
\code{get_wind()} will accept the output of 'make_extents()' and return a series of raster stacks, one for each time step, representing the desired wind related variables. The temporal resolution of the rasters will be the same as that
of the input wind extents.
}
\details{
Maximum sustained wind is defined differently across agencies and methods, but is generally summarised as the average wind speed sustained over a given period of time, the interval.
Intervals vary by agency, with 10-minutes being the most common and with USA agencies utilizing 1-minute intervals. In cyclones, wind speeds are highest towards the center and decrease
non-linearly as a function of the distance from the center. Several theoretical models have been proposed to generalize this relationship and can be utilized in 'get_wind()'. These functions
utilize the storm's location and its maximum wind speed and pressure differential to predict wind speeds anywhere within the footprint.
\href{"https://doi.org/10.1175/1520-0493(1980)108<1212:AAMOTW>2.0.CO;2"}{Holland, 1980} describes this as a function of the pressure field and the radius of maximum wind.
\href{"https://harvardforest.fas.harvard.edu/publications/pdfs/Boose_Chapter_HurricanesTyphoons_2004.pdf"}{Boose, 2004} expands upon this, introducing asymmetry and differential interactions
over land and water. \href{"https://doi.org/10.1175/MWR3106.1"}{Willoughby, 2006} is a piecewise function that stitches together two models, one for the inner region of the storm and another for
the outer region. The algorithms for these methods were adapted from the StormR package. Finally, an interpolation method, Thin Plate Spline, can be applied to the extents constructed by
'make_extents()'. Although computationally more demanding, TPS is shown here to best generalize observations from drop sondes.
}
\examples{
# One storm, coarse resolution, with native extents only and with consolidated values.
Helene <- get_storms(name='HELENE',season=2024)
Helene_extents <- make_extents(Helene)
Helene_winds <- get_wind(Helene_extents,methods='all')
Helene_winds <-rast(lapply(Helene_winds,unwrap))
Helene_msw <- Helene_winds[[grep("msw",names(Helene_winds))]]
plot(Helene_msw[[c(grep("tps",names(Helene_msw))[1],grep("boose",names(Helene_msw))[1],grep("holland",names(Helene_msw))[1],grep("willoughby",names(Helene_msw))[1])]])
##  take the maximum wind speed across the storm's lifetime
Helene_msw.max <- tapp(Helene_msw,c(ifelse(grepl("tps",names(Helene_msw)),"tps",ifelse(grepl("boose",names(Helene_msw)),"boose",ifelse(grepl("holland",names(Helene_msw)),"holland","willoughby")))),fun="max",na.rm=TRUE)
plot(Helene_msw.max)
}
