% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_extents.R
\name{make_extents}
\alias{make_extents}
\title{Create spatial simple features, linestrings or polygons, representing the storm track and wind extents of a storm at each timestep.}
\usage{
make_extents(
  storm,
  mods = NULL,
  type = "linestrings",
  t_res = NULL,
  agency = "CONS",
  cpus = NULL,
  todir = NULL
)
}
\arguments{
\item{storm}{a singular storm tabular data set as output by the \code{get_storms()} function. If a collection of storms is entered, only the first will be processed,
unless utilized in an lapply or parallel equivalent, in which case each will be processed individually.}

\item{mods}{the result of the \code{build_models()} function. Used to model necessary information when missing or erroneous, usually for older storms.
If absent, only native extents are produced and other rare missing or likely erroneous data are not adjusted. Examples of erroneous data include radii
of maximum wind that are greater than extents of lower wind speeds, or minimum central pressures that are greater than the pressure at the last closed isobar.}

\item{type}{the desired class of output sf object, either 'linestrings','polygons', or 'all'.}

\item{t_res}{the desired temporal resolution of the output features, in minutes.}

\item{cpus}{for internal parallelization. Currently not functional as parallel not faster for make_extents().}
}
\value{
Simple features geometry collection containing time stamped storm tracks and either linestrings, polygons, or both representing the maximum extent
of wind speeds, the eye wall (if present) and the Radius of the Last Closed Isobar (ROCI). All geometries are relative to a custom coordinate reference
system (crs) centered on the storm's centroid and in a Lambert azimuthal equal-area projection.
}
\description{
\code{make_extents()} will accept the output of \code{get_storms()} and create simple features geometry collections and data frames for each storm, with geo-located features
representing the maximum wind extent of various wind speeds at each time step of the storm, as well as the outer pressure extent (ROCI) and tracks representing the center of the storm.
By default, only the native time steps are included (usually every 3 hours), but the 't_res' argument can be used to change the temporal resolution and create
interpolated extents at different, usually more frequent, time steps. Unlike other functions in Cyclones, the 'make_extents()' iterations are relatively faster and computation times
typically increase, rather than decrease, in parallel. Thus, for now, parallel computation can only be performed at the storm level.
}
\examples{

storms <- get_storms()
##  will only compute the first storm
windextent <- make_extents(storms)
##  to compute all storms
windextents <- lapply(storms,make_extents)
library(ggplot2)
ggplot(windextent) + geom_sf(aes(col=kts))

# to include modeling, best to model based on entire dataset
allstorms <- get_storms(ib_filt="ALL")
mods <- build_models(allstorms)
windextent_mod <- make_extents(storms,mods=mods)
ggplot(windextent_mod) + geom_sf(aes(col=kts))

# use t_res to increase time step frequency
# here, every 30 mins
windextent_mod30 <- make_extents(storms,mods=mods,t_res=30)
ggplot(windextent_mod30) + geom_sf(aes(col=kts))

}
