% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precip.R
\name{get_precip}
\alias{get_precip}
\title{Generate 2D precipitation fields for a given storm}
\usage{
get_precip(
  storm,
  sources = "ecmwf",
  todir = NULL,
  dpath = NULL,
  t_res = NULL,
  s_res = NULL,
  prestorm = NULL,
  agg = FALSE,
  cpus = NULL,
  overwrite = FALSE,
  loadrasts = FALSE
)
}
\arguments{
\item{storm}{A set of storm tabular data generated by 'get_storms()' or wind extents generated by 'make_extents()'. If not used with an apply function, only the first storm will be processed.}

\item{sources}{Where to source the precipitation data, one or all of "ecmwf","mswep",or "gpm". "all" will return all. See details.}

\item{todir}{The target directory for final storm rasters. If not provided, rasters will be loaded in memory. Thus, highly recommended for high resolution products or for batch processing.}

\item{dpath}{The download path of the sourced precipitation data. Raw files will be stored here and then searched for here before downloading from the web. In some cases, these are global
files, meaning they can be used for any storm sharing the same date(s). For 'ecmwf' the files are cropped to storm extents before downloading and are this storm-specific.
Recommended for batch and repetitive analyses.}

\item{t_res}{The temporal resolution, in minutes, of the desired output. If not provided, the resolution of the source dataset will be used. If provided and different from the source data,
results will be interpolated to meet the desired resolution.}

\item{s_res}{The spatial resolution, in meters, of the desired output. Higher resolutions (lower numbers) will take significantly longer and parallel processing may be warranted below
10000 meters.}

\item{cpus}{How many cpu cores to assign for parallel processing. Recommended to verify available cores before using.}

\item{overwrite}{Should existing files in 'todir' be overwritten. If not, those will be skipped.}

\item{pre}{Numeric integer for the number of days of pre-storm precipitation to inlcude, as pre-cyclone precipitation events are acknowledged to sometimes heavily influence
soil moisture, surface flow, and storm-water dynamics during and after the storm.}
}
\description{
\code{get_precip()} will accept the output of either 'get_storms()' or 'make_extents()' and return a series of raster stacks, representing the total precipitation
accumulation for each time step. The temporal resolution of the rasters will be the same as that of the input data, unless otherwise specified. For all sources, some sort of
pre-registration for each source, and subsequent configuration in R is necessary to access data. See details. Additional packages are also necessary for API interaction at each source.
These will be prompted for when required. Alternatively, data could be downloaded manually and held in the './~dpath~/source/' directory, ensuring original filenames are retained,
for loading via 'get_precip()'.
}
\details{
It is generally accepted that 95\% of cyclone sourced precipitation falls within 500km of the center \href{"https://www.nature.com/articles/s41467-021-25685-2"}{Guzman & Jiang, 2012},
\href{"https://doi.org/10.1175\%2FJCLI-D-23-0666.1}{Chen et al. (2024)} (Future versions of this function will attempt to make the 500km distance variable). Under this premise,
precipitation datasets are subset by buffering a 500km radius around the storm's center at
each interval. Although time conventions differ by source, the product of 'get_precip()' will always reflect the end time for each interval, not the start time. "ecmwf" is default and is the
\href{"https://www.ecmwf.int/en/forecasts/dataset/ecmwf-reanalysis-v5"}{European Center for Medium-Range Weather Forecasts ERA5 dataset}, which has global coverage from 1940 to present
at hourly intervals and ~30 km spatial resolution. \href{"https://www.ecmwf.int/user/login}{Register for ECMWF here.} "mswep" is the
\href{"https://www.gloh2o.org/"}{Multi-Source Weighted-Ensemble Precipitation product} (V2) from GloH2O, which has global coverage from 1979 to present at 3-hourly intervals and 0.1 degree (10 km)
spatial resolution. These are housed in Google Drive and whichever account is used to request access to the data must also be the one used with 'googledrive::drive_auth()'.
Googleapis (the backend of the googledrive package) will sometimes block requests and downloads will fail. You may need to specifically clear the authorization cache and
re-grant Tidyverse API access to your account via 'unlink("~/.R/gargle/gargle-oauth")' and then 'googldrive::drive_auth(cache=FALSE)'. Otherwise, Google Drive does
implement throttling when download quotas are exceeded per file, regardless of who is requesting the file. In this case waiting is the only resolution. "gpm" is data from the
\href{"https://gpm.nasa.gov/missions/GPM"}{National Aeronautics and Space Administration's Global Precipitation Mission}, specifically the IMERG product, which is global coverage 30-minute interval
data from 2000 to the present. \href{"https://urs.earthdata.nasa.gov/"}{Register first here} and obtain a token, which then needs to be configured to R via 'earthdatalogin::edl_set_token()'.
Although MSWEP has the highest resolution and seems to be robust, it is proprietary and may eventually be inaccessible. Although the GPM data share the same
high resolution as MSWEP, the maximum values are much higher than the other sources and this overestimation \href{"https://journals.ametsoc.org/view/journals/hydr/24/10/JHM-D-22-0216.1.xml"}{has been noted}.
Thus, ecmwf is used by default.
}
\examples{
# One storm, coarse resolution, at native ~3hr intervals, all sources.
# first run will be relatively slow as data is acquired
# subsequent runs are much faster when 'dpath' is utilized or in the same R session
Helene <- get_storms(name='HELENE',season=2024)
Helene_extents <- make_extents(Helene)
Helene_precips <- get_precip(Helene_extents,sources='all',prestorm=TRUE)
##  precips separated into storm and pre-storm (if indicated) periods
Helene_prestorm_precips <- lapply(Helene_precips,function(x) x$prestorm)
Helene_precips <- lapply(Helene_precips,function(x) x$storm)
## original layer names retained within storm stacks
Helene_precips <-lapply(Helene_precips, function (x) unwrap(x))
Helene_prestorm_precips <-lapply(Helene_prestorm_precips, function (x) lapply(x,unwrap))
###  plot the same date across the different sources
plot(c(Helene_precips$HELENE_2024_NA_2024268N17278_precip_ecmwf[[50]],Helene_precips$HELENE_2024_NA_2024268N17278_precip_gpm[[50]],Helene_precips$HELENE_2024_NA_2024268N17278_precip_mswep[[50]]))
time(c(Helene_precips$HELENE_2024_NA_2024268N17278_precip_ecmwf[[50]],Helene_precips$HELENE_2024_NA_2024268N17278_precip_gpm[[50]],Helene_precips$HELENE_2024_NA_2024268N17278_precip_mswep[[50]]))
##  plot a date and all cells that fall within 48 hours before that time
Helene_t1 <- Helene_precips$HELENE_2024_NA_2024268N17278_precip_ecmwf[[50]]
time(Helene_t1)
Helene_t1_pre48hrs <- Helene_prestorm_precips$HELENE_2024_NA_2024268N17278_precip_ecmwf$prestorm_precip_48_hrs[[time(Helene_prestorm_precips$HELENE_2024_NA_2024268N17278_precip_ecmwf$prestorm_precip_48_hrs)==time(Helene_t1)-48*60*60]]
Helene_t1_minus48hrs <- Helene_precips$HELENE_2024_NA_2024268N17278_precip_ecmwf[[time(Helene_precips$HELENE_2024_NA_2024268N17278_precip_ecmwf)==time(Helene_precips$HELENE_2024_NA_2024268N17278_precip_ecmwf[[50]])-48*60*60]]
plot(c(Helene_t1,Helene_t1_pre48hrs, Helene_t1_minus48hrs))
##  take the sum of the precipitation across the storm's lifetime
Helene_precips.sum <- tapp(rast(Helene_precips),sapply(strsplit(names(rast(Helene_precips)),"_"),"[[",6),fun="sum",na.rm=TRUE)
plot(Helene_precips.sum,range=c(0,450))
## prestorm sum
Helene_prestorm48hrs.sum <- app(Helene_prestorm_precips$HELENE_2024_NA_2024268N17278_precip_ecmwf$prestorm_precip_48_hrs,fun="sum",na.rm=TRUE)
plot(Helene_prestorm48hrs.sum)
##  North Carolina received considerable pre-storm precipitation
plot(st_geometry(st_transform(rnaturalearth::ne_states(),crs(Helene_prestorm48hrs.sum))),add=TRUE,border="grey")

##  high resolution in parallel
Helene_precips <- get_precip(Helene_extents,sources='all',cpus=10,s_res=5000,t_res=15)
}
