% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storms.R
\name{get_storms}
\alias{get_storms}
\title{Gather storm(s) tabular data.}
\usage{
get_storms(
  source = "ncei",
  id = NULL,
  name = NULL,
  season = NULL,
  basin = NULL,
  ib_filt = NULL,
  consolidate = TRUE,
  cols = NULL,
  returndf = FALSE
)
}
\arguments{
\item{source}{string or data.frame input. If string, should be a target web source (either 'ncei' or 'hurdat') for downloading,
or a filename of a .csv or .nc file containing tabular cyclone data, or an already loaded data.frame. In the case of an existing data.frame, it should be consistent with
IBTrACS or HURDAT format.}

\item{id}{Either the SID (from IBTrACS) or the USA_ATCF_ID (from both hurdat and IBTrACS) used to identify individual storms. Results will be filtered to these values.}

\item{name}{The official name of the storm(s). Results will be filtered to these values.}

\item{season}{The year associated with the date of the storm(s). Results will be filtered to these values.}

\item{basin}{The oceanic basin of the storm(s). Results will be filtered to these values. In case a storm traverses multiple basins, its origin basin is used.}

\item{ib_filt}{A filtering value for downloading specific IBTrACS data and to avoid downloading the entire dataset if not required. Valid values are:
'ACTIVE', 'ALL', 'last3years', 'since1980', or the basins: "EP", "NA", "NI", "SA", "SI", "SP", "WP".}

\item{consolidate}{To run 'cons_stormdat()' or not, to consolidate values across agencies. When TRUE, CONS_ columns will be added to represent the consolidated values and, unless
the 'cols' parameter requests otherwise, the original columns will be dropped. If consolidation is not run, downstream functions may fail or return unexpected results,
as they may not be able to distinguish which columns to use. Also, for the maximum sustained wind columns, the interval period is inconsistent among agencies. Consolidation
translates these values to a common interval.}

\item{cols}{Which columns to retain in the output if consolidate is TRUE. Default is that only consolidated values are retained, in addition to storm identification, location,
and time information. Input values can be specific column names or agency prefixes, in which case all column from that agency are retained. If consolidation is FALSE,
all original columns are retained when 'cols' is empty.}
}
\value{
A list of data.frames, one for each storm, containing tabular information required for Cyclone processing. The names for each data.frame in the list are
unique identifiers for storms and used throughout Cyclones as filenames for various functions.
}
\description{
\code{get_storms()} will accept an already loaded data frame, a filename, or a target web source to download cyclone tabular data and
format it for processing in the Cyclones functions
}
\examples{
# default will get the last 3 years of storms from IBTrACS (ncei)
storms <- get_storms()

# get a specific storm
Maria_2017 <- get_storms(name="MARIA",season=2017)
Maria_2017 <- get_storms(id="2017260N12310")

# get all storms for a basin
NorAtl <- get_storms(basin="NA")

}
