\name{storms}
\alias{storms}
\alias{data(storms)}
\title{North Atlantic cyclone statistics.}
\description{Counts of low-pressure systems every 12hr over the North Atlantic region (00.00 and 12.00 UTC), derived from the NMC (now National Centers for Environmental Prediction, NCEP, \url{http://www.ncep.noaa.gov/}).}
\usage{data(storms)}
\format{'storm' - a list of lon, lat, psl, yy, mm, dd}
\source{The Norwegian Meteorological Institute, Climatology division.}
\references{EMS04-A-00146, Benestad, R.E.; Sorteberg, A.; Chen, D. 'Storm statistics derived using a calculus-based cyclone identification method', \url{http://www.cosis.net/members/meetings/sessions/oral_programme.php?p_id=110&s_id=1845}, European Meteorological Society AC2, Nice, Sept 28, 2004.}
\examples{
}
\keyword{datasets}
